pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pop_genome_checker"]

[project]
name = "pop-genome-checker"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.83",
    "click>=8.1",
    "pandas>=2.0",
    "rich>=13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[project.scripts]
pop-genome-checker = "pop_genome_checker.cli:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
